Parking models

+++ Vehicles/src/com/endava/models/contracts/Course.java
@@ -1,8 +1,10 @@
 package com.endava.models.contracts;
 
 import com.endava.models.vehicles.contracts.Vehicle;
+import java.math.BigDecimal;
 
 public interface Course {
+
   String getDestination();
 
   int getDistance();
@@ -11,6 +13,6 @@ public interface Course {
 
   Vehicle getVehicle();
 
-  double calculateTransportCosts();
+  BigDecimal calculateTransportCosts();
 
 }
