+++ Vehicles/src/com/endava/models/vehicles/contracts/Vehicle.java
@@ -1,11 +1,13 @@
 package com.endava.models.vehicles.contracts;
 
 import com.endava.models.vehicles.enums.VehicleType;
+import java.math.BigDecimal;
 
 public interface Vehicle {
+
   int getLoadCapacity();
 
-  double getPricePerKilometer();
+  BigDecimal getPricePerKilometer();
 
   VehicleType getType();
